backend/requirements.txt
fastapi
uvicorn
firebase-admin
apscheduler
google-generativeai
httpx
feedparser
xmltodict
pydantic-settings
python-dotenv
